Question/requirements.txt
Flask
pymysql
pandas
python-docx
PyPDF2
openpyxl
flask_sqlalchemy